# openapi_spec.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Carbon API",
        "version": "1.0.0",
        "description": "The Carbon API provides endpoints for user authentication, chat management, and messaging. Please Note: All parameters and responses are encrypted with RSA. Responses need to be decrypted with the private key pair of the public key used for registration. Parameters need to be encrypted with the public key used for authentication.",
        "contact": {
            "name": "Contributors to Carbon on GitHub",
            "url": "https://github.com/carbon-chat/carbon"
        }
    },
    "servers": [
        {
            "url": "https://carbon.proplayer919.dev:3000",
            "description": "API server"
        }
    ],
    "tags": [
        {
            "name": "Authentication",
            "description": "User authentication"
        },
        {
            "name": "Chat",
            "description": "Chat management"
        },
        {
            "name": "Health Check",
            "description": "Health check endpoint"
        },
        {
            "name": "Get Key",
            "description": "Get the server public key"
        }
    ],
    "paths": {
        "/api/v1/register": {
            "post": {
                "summary": "Register User",
                "description": "Registers a new user.",
                "operationId": "registerUser",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request if username or password is missing."
                    },
                    "409": {
                        "description": "Conflict if username already exists."
                    }
                }
            }
        },
        "/api/v1/auth": {
            "post": {
                "summary": "Authenticate User",
                "description": "Authenticates an existing user.",
                "operationId": "authenticateUser",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "username",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request if username or password is missing or incorrect."
                    },
                    "401": {
                        "description": "Unauthorized if authentication fails."
                    }
                }
            }
        },
        "/api/v1/updatePassword": {
            "post": {
                "summary": "Update Password",
                "description": "Updates the password for the authenticated user.",
                "operationId": "updatePassword",
                "tags": [
                    "Authentication"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully."
                    },
                    "400": {
                        "description": "Bad Request if password is missing."
                    }
                }
            }
        },
        "/api/v1/createChat": {
            "post": {
                "summary": "Create Chat",
                "description": "Creates a new chat room.",
                "operationId": "createChat",
                "tags": [
                    "Chat"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat room created successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "chatId": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request if name is missing."
                    },
                    "409": {
                        "description": "Conflict if a chat room with the same name already exists."
                    }
                }
            }
        },
        "/api/v1/createChatMessage": {
            "post": {
                "summary": "Create Chat Message",
                "description": "Adds a new message to a chat.",
                "operationId": "createChatMessage",
                "tags": [
                    "Chat"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "chatId": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "chatId",
                                    "content"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message added to chat successfully."
                    },
                    "400": {
                        "description": "Bad Request if chatId or content is missing."
                    },
                    "401": {
                        "description": "Unauthorized if user is not a member of the chat."
                    }
                }
            }
        },
        "/api/v1/getChatMessages": {
            "post": {
                "summary": "Get Chat Messages",
                "description": "Retrieves messages from a chat.",
                "operationId": "getChatMessages",
                "tags": [
                    "Chat"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "chatId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "chatId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Messages retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request if chatId is missing."
                    },
                    "401": {
                        "description": "Unauthorized if user is not a member of the chat."
                    }
                }
            }
        },
        "/api/v1/getInvolvedChats": {
            "post": {
                "summary": "Get Involved Chats",
                "description": "Retrieves IDs of chats involving the authenticated user.",
                "operationId": "getInvolvedChats",
                "tags": [
                    "Chat"
                ],
                "responses": {
                    "200": {
                        "description": "Chat IDs retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/getChatUsers": {
            "post": {
                "summary": "Get Chat Users",
                "description": "Retrieves users of a chat.",
                "operationId": "getChatUsers",
                "tags": [
                    "Chat"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "chatId": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "chatId"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request if chatId is missing."
                    },
                    "401": {
                        "description": "Unauthorized if user is not a member of the chat."
                    }
                }
            }
        },
        "/healthcheck": {
            "get": {
                "summary": "Health Check",
                "description": "Verifies server health.",
                "operationId": "healthCheck",
                "tags": [
                    "Health Check"
                ],
                "responses": {
                    "200": {
                        "description": "Server is healthy."
                    }
                }
            }
        },
        "/api/v1/getKey": {
            "get": {
                "summary": "Get Key",
                "description": "Retrieves the public key used for encryption.",
                "operationId": "getKey",
                "tags": [
                    "Get Key"
                ],
                "responses": {
                    "200": {
                        "description": "Public key retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}